Fix multi sample input

+++ automated.py
@@ -137,21 +137,27 @@ def process_input(src_dir, dest_dir, frame, sample_idx, suffix=None):
         spec_path = os.path.join(src_dir, f'specRough_{frame:04d}.exr')
         if os.path.exists(spec_path):
             spec_img = exr.read_all(spec_path)['default']
-            rough_img = spec_img[:,:,3:4]
-            exr.write(os.path.join(dest_dir, f'roughness_{frame:04d}.exr'), rough_img, compression=exr.ZIP_COMPRESSION)
-            exr.write(os.path.join(dest_dir, f'specularAlbedo_{frame:04d}.exr'), spec_img[:,:,0:3], compression=exr.ZIP_COMPRESSION)
-            os.remove(os.path.join(src_dir, f'specRough_{frame:04d}.exr'))
+            if spec_img.shape[-1] == 4:
+                rough_img = spec_img[:,:,3:4]
+                exr.write(os.path.join(dest_dir, f'roughness_{frame:04d}.exr'), rough_img, compression=exr.ZIP_COMPRESSION)
+                exr.write(os.path.join(dest_dir, f'specularAlbedo_{frame:04d}.exr'), spec_img[:,:,0:3], compression=exr.ZIP_COMPRESSION)
+                os.remove(os.path.join(src_dir, f'specRough_{frame:04d}.exr'))
+            else:
+                print("WARN: specRough has no alpha channel.")
         else:
             print(f'WARN: {spec_path} not found.')
 
         diffuseOpacity_path = os.path.join(src_dir, f'diffuseOpacity_{frame:04d}.exr')
         if os.path.exists(diffuseOpacity_path):
             diffuseOpacity_img = exr.read_all(diffuseOpacity_path)['default']
-            diffuse_img = diffuseOpacity_img[:,:,0:3]
-            opacity_img = diffuseOpacity_img[:,:,3:4]
-            exr.write(os.path.join(dest_dir, f'diffuseAlbedo_{frame:04d}.exr'), diffuse_img, compression=exr.ZIP_COMPRESSION)
-            exr.write(os.path.join(dest_dir, f'opacity_{frame:04d}.exr'), opacity_img, compression=exr.ZIP_COMPRESSION)
-            os.remove(os.path.join(src_dir, f'diffuseOpacity_{frame:04d}.exr'))
+            if diffuseOpacity_img.shape[-1] == 4:
+                diffuse_img = diffuseOpacity_img[:,:,0:3]
+                opacity_img = diffuseOpacity_img[:,:,3:4]
+                exr.write(os.path.join(dest_dir, f'diffuseAlbedo_{frame:04d}.exr'), diffuse_img, compression=exr.ZIP_COMPRESSION)
+                exr.write(os.path.join(dest_dir, f'opacity_{frame:04d}.exr'), opacity_img, compression=exr.ZIP_COMPRESSION)
+                os.remove(os.path.join(src_dir, f'diffuseOpacity_{frame:04d}.exr'))
+            else:
+                print("WARN: diffuseOpacity has no alpha channel.")
         else:
             print(f'WARN: {diffuseOpacity_path} not found.')
 
@@ -174,18 +180,17 @@ def process_input(src_dir, dest_dir, frame, sample_idx, suffix=None):
         rendered_files = os.listdir(dest_dir)
         rendered_files = [f for f in rendered_files if f.endswith(f'{frame:04d}.exr')]
         rendered_files = [f for f in rendered_files if not starts_with_number(f)]
+        rendered_files = [f for f in rendered_files if 'mvec' not in f]
 
         # Find the largest sample index from {sample_idx:04d}_*.exr
         if sample_idx == 0:
-            # Just move
+            # Just move, e.g., albedo_0100.exr -> 0000_albedo_0100.exr
             for f in rendered_files:
                 shutil.move(os.path.join(dest_dir, f), os.path.join(dest_dir, f'{sample_idx:04d}_{f}'))
         else:
             last_sample_idx = sample_idx - 1
             # Average the images
             for f in rendered_files:
-                if 'mvec' in f: # Do not average motion vectors
-                    continue
                 img_avg = exr.read_all(os.path.join(dest_dir, f'{last_sample_idx:04d}_{f}'))['default']
                 img = exr.read_all(os.path.join(dest_dir, f))['default']
                 new_avg = ((last_sample_idx+1) * img_avg + img) / ((last_sample_idx+1) + 1)
@@ -197,9 +202,16 @@ def process_input(src_dir, dest_dir, frame, sample_idx, suffix=None):
                 os.remove(os.path.join(dest_dir, f))
                 os.remove(os.path.join(dest_dir, f'{last_sample_idx:04d}_{f}'))
 
+        if sample_idx == config.SAMPLES_PER_PIXEL - 1:
+            # Rename the sample_idx from the filename
+            for f in rendered_files:
+                shutil.move(os.path.join(dest_dir, f'{sample_idx:04d}_{f}'), os.path.join(dest_dir, f))
+                print(f'Renaming {sample_idx:04d}_{f} to {f}')
+
     except Exception as e:
         func_name = sys._getframe()
-        print(f"[{func_name}] Error processing frame {frame}: {str(e)}")
+        line_number = sys.exc_info()[-1].tb_lineno
+        print(f"[{func_name}, line {line_number}] Error processing frame {frame}: {str(e)}")
 
 def postprocess_input(src_dir, scene_name, frames, sample_idx, suffix=None):
     print('\tPost-processing the input...', end=' ', flush=True)
